HU-17 added security check

+++ app/server/views/User_view.py
@@ -112,6 +112,17 @@ def modify_user(user_id: str, request: Request, data: UserSchema = Body(...)):
 
 @router.delete("/{user_id}")
 def remove_user(user_id: str, request: Request):
+    token_id = ""
+    if 'authorization' in request.headers:
+        try:
+            token_id = get_id_from_token(request.headers['authorization'].split(" ")[1])
+        except IndexError:
+            raise HTTPException(403, "Usuario incorrecto")
+
+    # Check if the user trying to watch the profile is the same
+    if token_id != user_id:
+        raise HTTPException(403, "Usuario incorrecto")
+
     removed = delete_user(user_id)
     if removed:
         return {"result": True}
